Update anndata.R - include support for Seurat V5
Updated the Seurat object access functions to support Assay5 Seurat objects.  Unsure if it is appropriate to utilize the object@assay$RNA@meta.data inplace of the meta.features from v4 due to the variances being split up.  Possibly need to run SCT v1 on this data to get the variances that match up with the Seurat V4 data object.